+++ api/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Bany API Change log
 
+## Version 0.6.0 - 17 Mar 2021
+
+**New features:**
+
+- API: User role: Administrator
+- API: API Key role: Admin
+- API: User role: Bany APP
+- API: API Key role: BanyApp
+
+- API: User role: Bany APP
+- API: GET Sync Artifact enndpoint
+- API: GET List Artifacts enndpoint
+- API: POST Upload Artifact enndpoint
+- API: DELETE Erase Artifact enndpoint
+
+- S3: Bany bucket
+
 ## Version 0.2.0 - 14 Mar 2021
 
 **New features:**